// admin/src/components/features/clients/detail/types.ts
export type ClientExtras = {
  nextApptLocation?: string | null;
  dietary?: string | null;
  allergies?: string | null;
  requests?: string | null;
  unwanted?: string | null;
  program?: string | null;
  email?: string | null;
  adults?: number | null;
  seniors?: number | null;
  children?: number | null;
  childrensAges?: string | null;
  itemsProvided?: string | null;
  notes?: string | null;
  dietaryRestrictions?: string | null;
  unwantedFoods?: string | null;
  additionalInfo?: string | null;
  hasMobilityIssues?: boolean | null;
  diaperSize?: string | null;
  ticketNumber?: string | null;
  nextAppointmentDate?: string | null;
  nextAppointmentTime?: string | null;
  completionTime?: string | null;
  checkInTime?: string | null;
  notificationPreference?: string | null;
  phoneCarrier?: string | null;
  location?: string | null;
  clientType?: string | null;
  provisions?: string | null;
  quantity?: number | null;
};

export type JoinedClient = {
  id: string;
  firstName: string;
  lastName: string;
  phone: string;
  householdSize: string;
  dietary: string | null;
  allergies: string | null;
  requests: string | null;
  unwanted: string | null;
  nextApptLocation: string | null;
  program: string | null;
  email: string | null;
  adults: number | null;
  seniors: number | null;
  children: number | null;
  childrensAges: string | null;
  itemsProvided: string | null;
  notes: string | null;
  dietaryRestrictions: string | null;
  unwantedFoods: string | null;
  additionalInfo: string | null;
  hasMobilityIssues: boolean | null;
  diaperSize: string | null;
  notificationPreference: string | null;
  phoneCarrier: string | null;
  location: string | null;
  clientType: string | null;
  provisions: string | null;
  quantity: number | null;
  ticketNumber: string | null;
  nextAppointmentDate: string | null;
  nextAppointmentTime: string | null;
  completionTime: string | null;
  checkInTime: string | null;
};